feat(vworker): impl process methods inside RebuildPlanner
