+++ venus-worker/src/sealing/worker/task/planner/common.rs
@@ -1,32 +1,65 @@
 //! this module provides some common handlers
 
 use std::collections::HashMap;
-use std::fs::{remove_file, File};
+use std::fs::{create_dir_all, remove_dir_all, remove_file, File, OpenOptions};
 use std::io::{self, prelude::*};
 use std::os::unix::fs::symlink;
+use std::path::{Path, PathBuf};
 
-use anyhow::{anyhow, Context};
+use anyhow::{anyhow, Context, Result};
 use vc_processors::builtin::tasks::STAGE_NAME_TREED;
 
-use super::super::{call_rpc, Entry, Task};
+use super::super::{call_rpc, cloned_required, field_required, Entry, Task};
 use crate::logging::debug;
-use crate::rpc::sealer::Deals;
+use crate::rpc::sealer::{Deals, SectorID, Seed, Ticket};
 use crate::sealing::failure::*;
 use crate::sealing::processor::{
-    cached_filenames_for_sector, tree_d_path_in_dir, write_and_preprocess, PieceInfo, RegisteredSealProof, TransferInput, TransferItem,
-    TransferRoute, TransferStoreInfo, TreeDInput, UnpaddedBytesAmount,
+    cached_filenames_for_sector, seal_commit_phase1, snap_generate_partition_proofs, snap_verify_sector_update_proof, tree_d_path_in_dir,
+    write_and_preprocess, PC1Input, PC2Input, PaddedBytesAmount, PieceInfo, RegisteredSealProof, SealCommitPhase1Output,
+    SealPreCommitPhase1Output, SealPreCommitPhase2Output, SnapEncodeInput, SnapEncodeOutput, SnapProveInput, SnapProveOutput,
+    TransferInput, TransferItem, TransferRoute, TransferStoreInfo, TreeDInput, UnpaddedBytesAmount, STAGE_NAME_C1, STAGE_NAME_PC1,
+    STAGE_NAME_PC2, STAGE_NAME_SNAP_ENCODE, STAGE_NAME_SNAP_PROVE,
 };
-use crate::types::SIZE_32G;
+use crate::sealing::util::get_all_zero_commitment;
+use crate::types::{SealProof, SIZE_32G};
+
+pub fn maybe_add_pieces<'t>(task: &'t Task<'_>, maybe_deals: Option<&Deals>) -> Result<Vec<PieceInfo>, Failure> {
+    let sector_id = task.sector_id()?;
+    let proof_type = task.sector_proof_type()?;
+
+    let mut staged_file = task.staged_file(sector_id).init_file().perm()?;
+
+    let seal_proof_type = proof_type.into();
+    let sector_size = proof_type.sector_size();
+    let mut pieces = Vec::new();
+
+    // acquired pieces
+    if let Some(deals) = maybe_deals {
+        add_pieces(task, seal_proof_type, &mut staged_file, &mut pieces, deals)?;
+    }
+
+    if pieces.is_empty() {
+        // skip AP for cc sector
+        staged_file.set_len(sector_size).context("add zero commitment").perm()?;
+
+        let commitment = get_all_zero_commitment(sector_size).context("get zero commitment").perm()?;
+
+        let unpadded_size: UnpaddedBytesAmount = PaddedBytesAmount(sector_size).into();
+        let pi = PieceInfo::new(commitment, unpadded_size).context("create piece info").perm()?;
+        pieces.push(pi);
+    }
+
+    Ok(pieces)
+}
 
 pub fn add_pieces<'t>(
     task: &'t Task<'_>,
     seal_proof_type: RegisteredSealProof,
     mut staged_file: &mut File,
+    pieces: &mut Vec<PieceInfo>,
     deals: &Deals,
-) -> Result<Vec<PieceInfo>, Failure> {
+) -> Result<(), Failure> {
     let piece_store = task.ctx.global.piece_store.as_ref().context("piece store is required").perm()?;
-
-    let mut pieces = Vec::new();
 
     for deal in deals {
         debug!(deal_id = deal.id, cid = %deal.piece.cid.0, payload_size = deal.payload_size, piece_size = deal.piece.size.0, "trying to add piece");
@@ -59,7 +92,7 @@ pub fn add_pieces<'t>(
         pieces.push(piece_info);
     }
 
-    Ok(pieces)
+    Ok(())
 }
 
 // build tree_d inside `prepare_dir` if necessary
@@ -102,6 +135,273 @@ pub fn build_tree_d(task: &'_ Task<'_>, allow_static: bool) -> Result<(), Failur
 
     drop(token);
     Ok(())
+}
+
+fn cleanup_before_pc1(cache_dir: &Path, sealed_file: &Path) -> Result<()> {
+    // TODO: see if we have more graceful ways to handle restarting pc1
+    remove_dir_all(&cache_dir).and_then(|_| create_dir_all(&cache_dir))?;
+    debug!("init cache dir {:?} before pc1", cache_dir);
+
+    let empty_sealed_file = OpenOptions::new()
+        .create(true)
+        .read(true)
+        .write(true)
+        .truncate(true)
+        .open(&sealed_file)?;
+
+    debug!("truncate sealed file {:?} before pc1", sealed_file);
+    drop(empty_sealed_file);
+
+    Ok(())
+}
+
+pub fn pre_commit1(task: &'_ Task<'_>) -> Result<(Ticket, SealPreCommitPhase1Output), Failure> {
+    let token = task.ctx.global.limit.acquire(STAGE_NAME_PC1).crit()?;
+
+    let sector_id = task.sector_id()?;
+    let proof_type = task.sector_proof_type()?;
+
+    let ticket = call_rpc! {
+        task.ctx.global.rpc,
+        assign_ticket,
+        sector_id.clone(),
+    }?;
+
+    debug!(ticket = ?ticket.ticket.0, epoch = ticket.epoch, "ticket assigned from sector-manager");
+
+    field_required! {
+        piece_infos,
+        task.sector.phases.pieces.as_ref().cloned()
+    }
+
+    let cache_dir = task.cache_dir(sector_id);
+    let staged_file = task.staged_file(sector_id);
+    let sealed_file = task.sealed_file(sector_id);
+    let prepared_dir = task.prepared_dir(sector_id);
+
+    cleanup_before_pc1(cache_dir.as_ref(), sealed_file.as_ref()).crit()?;
+    symlink(tree_d_path_in_dir(prepared_dir.as_ref()), tree_d_path_in_dir(cache_dir.as_ref())).crit()?;
+
+    field_required! {
+        prove_input,
+        task.sector.base.as_ref().map(|b| b.prove_input)
+    }
+
+    let out = task
+        .ctx
+        .global
+        .processors
+        .pc1
+        .process(PC1Input {
+            registered_proof: (*proof_type).into(),
+            cache_path: cache_dir.into(),
+            in_path: staged_file.into(),
+            out_path: sealed_file.into(),
+            prover_id: prove_input.0,
+            sector_id: prove_input.1,
+            ticket: ticket.ticket.0,
+            piece_infos,
+        })
+        .perm()?;
+
+    drop(token);
+    Ok((ticket, out))
+}
+
+fn cleanup_before_pc2(cache_dir: &Path) -> Result<()> {
+    for entry_res in cache_dir.read_dir()? {
+        let entry = entry_res?;
+        let fname = entry.file_name();
+        if let Some(fname_str) = fname.to_str() {
+            let should = fname_str == "p_aux" || fname_str == "t_aux" || fname_str.contains("tree-c") || fname_str.contains("tree-r-last");
+
+            if !should {
+                continue;
+            }
+
+            let p = entry.path();
+            remove_file(&p).with_context(|| format!("remove cached file {:?}", p))?;
+            debug!("remove cached file {:?} before pc2", p);
+        }
+    }
+
+    Ok(())
+}
+
+pub fn pre_commit2(task: &'_ Task) -> Result<SealPreCommitPhase2Output, Failure> {
+    let token = task.ctx.global.limit.acquire(STAGE_NAME_PC2).crit()?;
+
+    let sector_id = task.sector_id()?;
+
+    field_required! {
+        pc1out,
+        task.sector.phases.pc1out.as_ref().cloned()
+    }
+
+    let cache_dir = task.cache_dir(sector_id);
+    let sealed_file = task.sealed_file(sector_id);
+
+    cleanup_before_pc2(cache_dir.as_ref()).crit()?;
+
+    let out = task
+        .ctx
+        .global
+        .processors
+        .pc2
+        .process(PC2Input {
+            pc1out,
+            cache_dir: cache_dir.into(),
+            sealed_file: sealed_file.into(),
+        })
+        .perm()?;
+
+    drop(token);
+    Ok(out)
+}
+
+pub fn commit1_with_seed(task: &Task, seed: Seed) -> Result<SealCommitPhase1Output, Failure> {
+    let token = task.ctx.global.limit.acquire(STAGE_NAME_C1).crit()?;
+
+    let sector_id = task.sector_id()?;
+
+    field_required! {
+        prove_input,
+        task.sector.base.as_ref().map(|b| b.prove_input)
+    }
+
+    let (seal_prover_id, seal_sector_id) = prove_input;
+
+    field_required! {
+        piece_infos,
+        task.sector.phases.pieces.as_ref()
+    }
+
+    field_required! {
+        ticket,
+        task.sector.phases.ticket.as_ref()
+    }
+
+    cloned_required! {
+        p2out,
+        task.sector.phases.pc2out
+    }
+
+    let cache_dir = task.cache_dir(sector_id);
+    let sealed_file = task.sealed_file(sector_id);
+
+    let out = seal_commit_phase1(
+        cache_dir.into(),
+        sealed_file.into(),
+        seal_prover_id,
+        seal_sector_id,
+        ticket.ticket.0,
+        seed.seed.0,
+        p2out,
+        piece_infos,
+    )
+    .perm()?;
+
+    drop(token);
+    Ok(out)
+}
+
+pub fn snap_encode(task: &Task, sector_id: &SectorID, proof_type: &SealProof) -> Result<SnapEncodeOutput, Failure> {
+    let _token = task.ctx.global.limit.acquire(STAGE_NAME_SNAP_ENCODE).crit()?;
+
+    cloned_required!(piece_infos, task.sector.phases.pieces);
+
+    // sealed file & cache dir should be ready now, we just do nothing here
+    let sealed_file = task.sealed_file(sector_id);
+    let cache_dir = task.cache_dir(sector_id);
+
+    // init update file
+    let update_file = task.update_file(sector_id);
+    debug!(path=?update_file.full(),  "trying to init update file");
+    {
+        let file = update_file.init_file().perm()?;
+        file.set_len(proof_type.sector_size()).context("fallocate for update file").perm()?;
+    }
+
+    let update_cache_dir = task.update_cache_dir(sector_id);
+    debug!(path=?update_cache_dir.full(),  "trying to init update cache dir");
+    update_cache_dir.prepare().context("prepare update cache dir").perm()?;
+
+    // tree d
+    debug!("trying to prepare tree_d");
+    let prepared_dir = task.prepared_dir(sector_id);
+    symlink(
+        tree_d_path_in_dir(prepared_dir.as_ref()),
+        tree_d_path_in_dir(update_cache_dir.as_ref()),
+    )
+    .context("link prepared tree_d")
+    .crit()?;
+
+    // staged file should be already exists, do nothing
+    let staged_file = task.staged_file(sector_id);
+
+    task.ctx
+        .global
+        .processors
+        .snap_encode
+        .process(SnapEncodeInput {
+            registered_proof: proof_type.into(),
+            new_replica_path: update_file.into(),
+            new_cache_path: update_cache_dir.into(),
+            sector_path: sealed_file.into(),
+            sector_cache_path: cache_dir.into(),
+            staged_data_path: staged_file.into(),
+            piece_infos,
+        })
+        .perm()
+}
+
+pub fn snap_prove(task: &Task) -> Result<SnapProveOutput, Failure> {
+    let _token = task.ctx.global.limit.acquire(STAGE_NAME_SNAP_PROVE).crit()?;
+
+    let sector_id = task.sector_id()?;
+    let proof_type = task.sector_proof_type()?;
+    field_required!(encode_out, task.sector.phases.snap_encode_out.as_ref());
+    field_required!(comm_r_old, task.sector.finalized.as_ref().map(|f| f.public.comm_r));
+
+    let sealed_file = task.sealed_file(sector_id);
+    let cached_dir = task.cache_dir(sector_id);
+    let update_file = task.update_file(sector_id);
+    let update_cache_dir = task.update_cache_dir(sector_id);
+
+    let vannilla_proofs = snap_generate_partition_proofs(
+        (*proof_type).into(),
+        comm_r_old,
+        encode_out.comm_r_new,
+        encode_out.comm_d_new,
+        sealed_file.into(),
+        cached_dir.into(),
+        update_file.into(),
+        update_cache_dir.into(),
+    )
+    .perm()?;
+
+    let proof = task
+        .ctx
+        .global
+        .processors
+        .snap_prove
+        .process(SnapProveInput {
+            registered_proof: (*proof_type).into(),
+            vannilla_proofs: vannilla_proofs.into_iter().map(|b| b.0).collect(),
+            comm_r_old,
+            comm_r_new: encode_out.comm_r_new,
+            comm_d_new: encode_out.comm_d_new,
+        })
+        .perm()?;
+
+    let verified =
+        snap_verify_sector_update_proof(proof_type.into(), &proof, comm_r_old, encode_out.comm_r_new, encode_out.comm_d_new).perm()?;
+
+    if !verified {
+        return Err(anyhow!("generated an invalid update proof").perm());
+    }
+
+    Ok(proof)
 }
 
 // acquire a persist store for sector files, copy the files and return the instance name of the
