+++ venus-worker/src/sealing/worker/task/planner/rebuild.rs
@@ -1,12 +1,13 @@
-use anyhow::Result;
+use anyhow::{Context, Result};
 
 use super::{
-    super::{call_rpc, Event, State, Task},
-    ExecResult, Planner,
+    super::{call_rpc, field_required, Event, State, Task},
+    common, ExecResult, Planner,
 };
 use crate::logging::warn;
-use crate::rpc::sealer::AllocateSectorSpec;
+use crate::rpc::sealer::{AllocateSectorSpec, Seed};
 use crate::sealing::failure::*;
+use crate::sealing::processor::STAGE_NAME_SNAP_ENCODE;
 
 pub struct RebuildPlanner;
 
@@ -25,6 +26,10 @@ struct Rebuild<'c, 't> {
 }
 
 impl<'c, 't> Rebuild<'c, 't> {
+    fn is_snapup(&self) -> bool {
+        self.task.sector.finalized.is_some()
+    }
+
     fn empty(&self) -> ExecResult {
         let maybe_res = call_rpc! {
             self.task.ctx.global.rpc,
@@ -55,42 +60,78 @@ impl<'c, 't> Rebuild<'c, 't> {
     }
 
     fn add_pieces_for_sealing(&self) -> ExecResult {
-        unimplemented!()
+        // if this is a snapup sector, then the deals should be used later
+        let maybe_deals = if self.is_snapup() { None } else { self.task.sector.deals.as_ref() };
+
+        let pieces = common::maybe_add_pieces(self.task, maybe_deals)?;
+
+        Ok(Event::AddPiece(pieces))
     }
 
     fn build_tree_d_for_sealing(&self) -> ExecResult {
-        unimplemented!()
+        common::build_tree_d(self.task, true)?;
+        Ok(Event::BuildTreeD)
     }
 
     fn pc1(&self) -> ExecResult {
-        unimplemented!()
+        let (ticket, out) = common::pre_commit1(self.task)?;
+        Ok(Event::PC1(ticket, out))
     }
 
     fn pc2(&self) -> ExecResult {
-        unimplemented!()
+        common::pre_commit2(self.task).map(Event::PC2)
     }
 
     fn check_sealed(&self) -> ExecResult {
-        unimplemented!()
+        field_required! {
+            ticket,
+            self.task.sector.phases.ticket.as_ref()
+        }
+
+        let seed = Seed {
+            seed: ticket.ticket,
+            epoch: ticket.epoch,
+        };
+
+        common::commit1_with_seed(self.task, seed).map(|_| Event::CheckSealed)
     }
 
     fn add_piece_for_snapup(&self) -> ExecResult {
-        unimplemented!()
+        if !self.is_snapup() {
+            return Ok(Event::SkipSnap);
+        }
+
+        field_required!(deals, self.task.sector.deals.as_ref());
+
+        common::maybe_add_pieces(self.task, Some(deals)).map(Event::AddPiece)
     }
 
     fn build_tree_d_for_snapup(&self) -> ExecResult {
-        unimplemented!()
+        common::build_tree_d(self.task, false).map(|_| Event::BuildTreeD)
     }
 
     fn snap_encode(&self) -> ExecResult {
-        unimplemented!()
+        let sector_id = self.task.sector_id()?;
+        let proof_type = self.task.sector_proof_type()?;
+
+        common::snap_encode(self.task, sector_id, proof_type).map(Event::SnapEncode)
     }
 
     fn snap_prove(&self) -> ExecResult {
-        unimplemented!()
+        common::snap_prove(self.task).map(Event::SnapProve)
     }
 
     fn persist(&self) -> ExecResult {
-        unimplemented!()
+        let sector_id = self.task.sector_id()?;
+
+        let (cache_dir, sealed_file) = if self.is_snapup() {
+            (self.task.update_cache_dir(sector_id), self.task.update_file(sector_id))
+        } else {
+            (self.task.cache_dir(sector_id), self.task.sealed_file(sector_id))
+        };
+
+        common::persist_sector_files(self.task, cache_dir, sealed_file).map(Event::Persist)
     }
+
+    fn submit_persist(&self) {}
 }
